moved the spinner code to the console
